Pterodactyl Support
This is the Pterodactyl multi-egg support, you can disable this if you aren't using multi-egg
This allows "stop" command that acts exactly as "quit" command. Also sets the server to "Online" state by sending the message "Done (38.841s)! For help, type "help" or "?""
Also this uses the multi-egg base to Minecraft, this means why the message and the command